requirements.txt
tqdm
Telethon
Pillow